pyproject.toml
[tool.poetry]
name = "she-dhos-janitor-api"
version = "0.1.0"
description = "DHOS Janitor API"
authors = ["Ian Simpson <ian.simpson@sensynehealth.com>"]
keywords = ["Swagger", "Janitor", "Reset"]
repository = "https://github.com/draysontechnologies/dhos-janitor-api"

[[tool.poetry.source]]
name = "sensynehealth"
url = "https://pypi.fury.io/sensynehealth/"

[tool.poetry.dependencies]
python = "^3.9,<3.10"
auth0-api-client = "1.*"
cachetools = "5.*"
draymed = "2.*"
faker = "4.*"
flask-batteries-included = {version = "3.*", extras = ["apispec"]}
httpx = "0.*"
python-jose = "3.*"
she-data-generation = "1.*"
she-logging = "1.*"

[tool.poetry.dev-dependencies]
bandit = "*"
black = "*"
coloredlogs = "*"
coverage = "*"
isort = "*"
mock = "*"
mypy = "*"
pytest = "*"
pytest-dhos = "*"
pytest-flask = "*"
pytest-mock = "*"
respx = "0.*"
safety = "*"
tox = "*"
types-cachetools = "*"
types-Flask = "*"
types-mock = "*"
types-PyYAML = "*"
types-requests = "*"
types-waitress = "*"

[tool.mypy]
namespace_packages=true
ignore_missing_imports=false
disallow_untyped_defs=true

[[tool.mypy.overrides]]
module = [
    "jose.*",
    "apispec.*",
    "apispec_webframeworks.*",
    "connexion",
    "dhosredis",
    "jsonschema",
    "faker"
]
ignore_missing_imports = true

[tool.isort]
profile = "black"
known_third_party = ["apispec", "apispec_webframeworks", "auth0_api_client", "behave", "cachetools", "click", "clients", "connexion", "draymed", "environs", "faker", "flask", "flask_batteries_included", "helpers", "httpx", "jose", "marshmallow", "mock", "pytest", "pytest_mock", "reporting", "reportportal_behave", "requests", "respx", "she_data_generation", "she_logging", "waitress", "yaml"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.pytest.ini_options]
filterwarnings = "ignore::DeprecationWarning:marshmallow.fields"
